refactor(config): split into multiple source files under rime/config
