+++ src/CMakeLists.txt
@@ -3,6 +3,7 @@ set(LIBRARY_OUTPUT_PATH ${PROJECT_BINARY_DIR}/lib)
 aux_source_directory(. rime_src_api)
 aux_source_directory(rime rime_src_base)
 aux_source_directory(rime/algo rime_src_algo)
+aux_source_directory(rime/config rime_src_config)
 aux_source_directory(rime/dict rime_src_dict)
 aux_source_directory(rime/gear rime_src_gear)
 aux_source_directory(rime/lever rime_src_lever)
